requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyside6
qrcode[pil]
pillow
python-multipart
jinja2